jzdz/requirements/project.txt
-f http://shrub.appspot.com/jzdz/
--extra-index-url=http://dist.gondor.io/dev/

django-ittybitty==0.1.0-pre3
Django==1.2.5
django-dbtemplates

nashvegas==0.5.2
gondor==1.0a1.dev20
psycopg2==2.3.2
django-staticfiles==1.0.1
